Refactor command bus configuration, split between tweet and infrastructure.
