+++ src/main/java/com/scmspain/MsFcTechTestApplication.java
@@ -1,6 +1,5 @@
 package com.scmspain;
 
-import com.scmspain.infrastructure.configuration.CommandBusConfiguration;
 import com.scmspain.infrastructure.configuration.InfrastructureConfiguration;
 import com.scmspain.infrastructure.configuration.TweetConfiguration;
 import org.springframework.boot.SpringApplication;
@@ -10,7 +9,7 @@ import org.springframework.context.annotation.Import;
 
 @Configuration
 @EnableAutoConfiguration
-@Import({TweetConfiguration.class, InfrastructureConfiguration.class, CommandBusConfiguration.class})
+@Import({TweetConfiguration.class, InfrastructureConfiguration.class})
 public class MsFcTechTestApplication {
 
     public static void main(String[] args) {
