+++ src/main/java/com/scmspain/infrastructure/configuration/TweetConfiguration.java
@@ -5,6 +5,8 @@ import com.scmspain.application.services.TweetMetricService;
 import com.scmspain.application.services.TweetValidationService;
 import com.scmspain.domain.TweetService;
 import com.scmspain.domain.command.CommandBus;
+import com.scmspain.domain.command.ListAllTweetsCommandHandler;
+import com.scmspain.domain.command.PublishTweetCommandHandler;
 import com.scmspain.infrastructure.controller.TweetController;
 
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -18,14 +20,24 @@ import org.springframework.context.annotation.Configuration;
 public class TweetConfiguration {
 
     @Bean("mainTweetService")
-    public TweetService getTweetService(@Qualifier("tweetRepository") TweetService tweetService, MetricService metricService) {
+    public TweetService getTweetService(final @Qualifier("tweetRepository") TweetService tweetService, final MetricService metricService) {
         TweetService tweetValidationService = new TweetValidationService(tweetService);
         return new TweetMetricService(tweetValidationService, metricService);
     }
 
     @Bean
-    public TweetController getTweetConfiguration(final CommandBus commandBus) {
+    public TweetController getTweetController(final CommandBus commandBus) {
         return new TweetController(commandBus);
+    }
+
+    @Bean
+    public PublishTweetCommandHandler getPublishTweetCommandHandler(@Qualifier("mainTweetService") final TweetService tweetService) {
+        return new PublishTweetCommandHandler(tweetService);
+    }
+
+    @Bean
+    public ListAllTweetsCommandHandler getListAllTweetsCommandHandler(@Qualifier("mainTweetService") final TweetService tweetService) {
+        return new ListAllTweetsCommandHandler(tweetService);
     }
 
 }
